Replace custom RNG with cpp std

See merge request Scribble/gdlenhanced!325

+++ Source/HotSpot.cpp
@@ -99,7 +99,7 @@ void CHotSpotWeenie::DoCycleDamage(CWeenieObject *other)
 	DamageEventData damageEvent;
 	damageEvent.damage_type = (DAMAGE_TYPE) m_Qualities.GetInt(DAMAGE_TYPE_INT, 0);
 	damageEvent.damage_form = DAMAGE_FORM::DF_HOTSPOT;
-	damageEvent.damageAfterMitigation = damageEvent.damageBeforeMitigation = (int)(m_Qualities.GetInt(DAMAGE_INT, 0) * (1.0 - (Random::GenFloat(0.0f, m_Qualities.GetFloat(DAMAGE_VARIANCE_FLOAT, 0.0)))));
+	damageEvent.damageAfterMitigation = damageEvent.damageBeforeMitigation = (int)(m_Qualities.GetInt(DAMAGE_INT, 0) * (1.0 - (Random::GenFloat(0.0f, (float)m_Qualities.GetFloat(DAMAGE_VARIANCE_FLOAT, 0.0)))));
 	damageEvent.target = other;
 	damageEvent.source = this;
